print python error in java
