+++ java/fury-core/src/test/java/org/apache/fury/CrossLanguageTest.java
@@ -48,7 +48,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
@@ -66,6 +65,7 @@ import org.apache.fury.serializer.BufferObject;
 import org.apache.fury.serializer.EnumSerializerTest;
 import org.apache.fury.serializer.Serializer;
 import org.apache.fury.serializer.StructSerializer;
+import org.apache.fury.test.TestUtils;
 import org.apache.fury.util.DateTimeUtils;
 import org.apache.fury.util.MurmurHash3;
 import org.testng.Assert;
@@ -84,27 +84,8 @@ public class CrossLanguageTest extends FuryTestBase {
    * @return Whether the command succeeded.
    */
   private boolean executeCommand(List<String> command, int waitTimeoutSeconds) {
-    return executeCommand(
+    return TestUtils.executeCommand(
         command, waitTimeoutSeconds, ImmutableMap.of("ENABLE_CROSS_LANGUAGE_TESTS", "true"));
-  }
-
-  private boolean executeCommand(
-      List<String> command, int waitTimeoutSeconds, Map<String, String> env) {
-    try {
-      LOG.info("Executing command: {}", String.join(" ", command));
-      ProcessBuilder processBuilder =
-          new ProcessBuilder(command)
-              .redirectOutput(ProcessBuilder.Redirect.INHERIT)
-              .redirectError(ProcessBuilder.Redirect.INHERIT);
-      for (Map.Entry<String, String> entry : env.entrySet()) {
-        processBuilder.environment().put(entry.getKey(), entry.getValue());
-      }
-      Process process = processBuilder.start();
-      process.waitFor(waitTimeoutSeconds, TimeUnit.SECONDS);
-      return process.exitValue() == 0;
-    } catch (Exception e) {
-      throw new RuntimeException("Error executing command " + String.join(" ", command), e);
-    }
   }
 
   @Data
