+++ java/fury-format/src/test/java/org/apache/fury/format/CrossLanguageTest.java
@@ -25,9 +25,7 @@ import static org.apache.fury.format.vectorized.ArrowUtilsTest.createVectorSchem
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.nio.channels.Channels;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -38,7 +36,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import lombok.Data;
 import org.apache.arrow.memory.BufferAllocator;
@@ -70,6 +67,7 @@ import org.apache.fury.logging.LoggerFactory;
 import org.apache.fury.memory.MemoryBuffer;
 import org.apache.fury.memory.MemoryUtils;
 import org.apache.fury.serializer.BufferObject;
+import org.apache.fury.test.TestUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -341,41 +339,8 @@ public class CrossLanguageTest {
    * @return Whether the command succeeded.
    */
   private boolean executeCommand(List<String> command, int waitTimeoutSeconds) {
-    return executeCommand(
+    return TestUtils.executeCommand(
         command, waitTimeoutSeconds, ImmutableMap.of("ENABLE_CROSS_LANGUAGE_TESTS", "true"));
-  }
-
-  private boolean executeCommand(
-      List<String> command, int waitTimeoutSeconds, Map<String, String> env) {
-    try {
-      LOG.info("Executing command: {}", String.join(" ", command));
-      // redirectOutput doesn't work for forked jvm such as in maven sure.
-      ProcessBuilder processBuilder = new ProcessBuilder(command);
-      for (Map.Entry<String, String> entry : env.entrySet()) {
-        processBuilder.environment().put(entry.getKey(), entry.getValue());
-      }
-      Process process = processBuilder.start();
-      // Capture output to log
-      BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-      BufferedReader errorReader =
-          new BufferedReader(new InputStreamReader(process.getErrorStream()));
-      String line;
-      while ((line = reader.readLine()) != null) {
-        System.out.println(line);
-      }
-      while ((line = errorReader.readLine()) != null) {
-        System.err.println(line);
-      }
-      boolean finished = process.waitFor(waitTimeoutSeconds, TimeUnit.SECONDS);
-      if (finished) {
-        return process.exitValue() == 0;
-      } else {
-        process.destroy(); // ensure the process is terminated
-        return false;
-      }
-    } catch (Exception e) {
-      throw new RuntimeException("Error executing command " + String.join(" ", command), e);
-    }
   }
 
   @Test
